2006-07-06  Martin Baulig  <martin@ximian.com>
	* backends/server/i386-arch.c (ArchInfo): Use `guint64' for
	`dr_control' and `dr_status' since _server_ptrace_get_dr()
	operates on guint64.


svn path=/trunk/debugger/; revision=62295

+++ backends/server/i386-arch.c
@@ -23,7 +23,7 @@ struct ArchInfo
 	INFERIOR_FPREGS_TYPE *saved_fpregs;
 	int saved_signal;
 	GPtrArray *rti_stack;
-	unsigned dr_control, dr_status;
+	guint64 dr_control, dr_status;
 	int dr_regs [DR_NADDR];
 };
